+++ hack/runtime-migrator/internal/init/client.go
@@ -1,4 +1,4 @@
-package input
+package init
 
 import (
 	"fmt"
