+++ hack/runtime-migrator/internal/init/migrationparams.go
@@ -1,4 +1,4 @@
-package input
+package init
 
 import (
 	"bufio"
